Use a single STL mesh for cabinet collision checking
The DAE file actually has 4 meshes inside: 2 each for the main cabinet
and the little box jutting out of it at the top right, and in each case
one for the door and the other for the other sides. I don't know how
Gazebo deals with meshes like this for collision detection, but
replacing the whole thing with a single STL mesh seems to lead to more
stable physics simulation.